% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title Binomial experiment}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
A bar plot
}
\description{
Title Binomial experiment
}
\examples{
mybin(100, 10, 0.7)
}
