% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{myboot function}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample size}

\item{fun}{desired statistic}

\item{alpha}{interval}

\item{...}{}
}
\description{
myboot function
}
\examples{
\dontrun{myboot(10000,x=sam1,fun="mean",alpha=0.05,xlab="mean",col="Purple")}
}
