% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Title uniform distribution}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{number of rows}

\item{iter}{number of columns}
}
\value{
a histogram
}
\description{
Title uniform distribution
}
\examples{
myclt(10,1000)
}
