% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Title mymlbeta}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{values}

\item{alpha}{shape 1}

\item{beta}{shape 2}

\item{...}{}
}
\description{
Title mymlbeta
}
\examples{
mymlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,4,length=100),beta=seq(2,8,length=100),lwd=2,labcex=1)
}
