% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{myncurve}
\alias{myncurve}
\title{Title a normal curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{endpoint}
}
\value{
a normal graph
}
\description{
Title a normal curve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
