% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myplot}
\alias{myplot}
\title{A function that returns a quadratic expression}
\usage{
myplot(x)
}
\arguments{
\item{x}{The root of the quadratic function}
}
\value{
A quadratic expression
}
\description{
A function that returns a quadratic expression
}
\examples{
myplot(x = 2)
}
