% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title Number of tickets to sell determiner}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of available sits on the plane}

\item{gamma}{The probability that the plane will be overbooked}

\item{p}{The probability of a show}
}
\value{
2 graphs and a list
}
\description{
Title Number of tickets to sell determiner
}
\examples{
ntickets(400,0.02,0.95)
}
